some tidying first
