+++ .pylintrc
@@ -2,7 +2,8 @@
 indent-string='  '
 
 [MESSAGES CONTROL]
-disable=C0114,C0116,W0311
+# R0903 - Too many instance attributes
+disable=C0114,C0116,W0311,R0902,R0903
 
 [FORMAT]
 indent-string='  '
