+++ app/utils/common.py
@@ -1,0 +1,13 @@
+def get_menu_items():
+    return [
+        {
+          'link': 'https://github.com/xarillian',
+          'display': 'github',
+          'icon': 'fab fa-github'
+        },
+        {
+          'link': 'https://www.linkedin.com/in/austin-heinrich/',
+          'display': 'linkedin',
+          'icon': 'fab fa-linkedin'
+        },
+    ]
